+++ MostAwesomeDartApplicationEver/MostAwesomeDartApplicationEver/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="DartScore.App"
+﻿<Application x:Class="MostAwesomeDartApplicationEver.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:DartScore">
+             xmlns:local="clr-namespace:MostAwesomeDartApplicationEver">
     <Application.Resources>
          
     </Application.Resources>
