+++ MostAwesomeDartApplicationEver/MostAwesomeDartApplicationEver/Models/Darter.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DartScore.Models
+namespace MostAwesomeDartApplicationEver.Models
 {
     internal class Darter
     {
